requirements.txt
asttokens==2.0.8
backcall==0.2.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
executing==1.1.1
Flask==2.2.2
Flask-MySQLdb==1.0.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mysqlclient==2.1.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
pure-eval==0.2.2
Pygments==2.13.0
python-slugify==6.1.2
six==1.16.0
stack-data==0.5.1
text-unidecode==1.3
traitlets==5.4.0
wcwidth==0.2.5
Werkzeug==2.2.2
